用户端扫描功能基本实现，bug修改进行中，已完成分享功能bug修改

+++ YKY/yaogouCell.h
@@ -22,7 +22,7 @@
 
 @property (weak, nonatomic) IBOutlet UILabel *plimitLabel;
 
-@property (nonatomic ) int index;
+@property (nonatomic ) long index;
 
 @property (nonatomic , strong) homeNewScuessModel * prizeModel;
